requirements.txt
# Core ML
torch
torchvision
torchaudio
numpy

# Data processing
pyyaml
pandas

# Visualization  
matplotlib
plotly

# Video generation
pillow